I added the auth service for the auth controller
