+++ src/controllers/auth.controller.ts
@@ -4,24 +4,13 @@ import { match } from '../utils/hashing';
 import { Users } from '../entity/users';
 import Jwt from "jsonwebtoken";
 import Boom from '@hapi/boom';
+import * as AuthService from '../services/auth';
 
 export const register = async (req: Request, res: Response, next: NextFunction) => {
     try {
-        const userRepository = getRepository(Users);
         const { name, lastname, email, password } = req.body;
-        const existUser = await userRepository.findOne({'email': email});
-        if (!existUser) {
-            var newUser = {
-                name,
-                lastname,
-                email,
-                password
-            }
-            const userData = userRepository.create(newUser);
-            const results = await userRepository.save(userData);
-            return res.json(results);
-        }
-        return res.json({msg: 'Username already exists!'});
+        const newUser = await AuthService.create(name, lastname, email, password);
+        return res.json(newUser);
     } catch (err) {
         return next(Boom.badRequest(err.message));
     }
@@ -30,25 +19,8 @@ export const register = async (req: Request, res: Response, next: NextFunction) 
 export const login = async (req: Request, res: Response , next: NextFunction) => {
     try {
         const { email, password} = req.body;
-        const user = await getRepository(Users).findOne({'email': email});
-        if(user){
-            let isPasswordMatching = await match(password, user.password);
-            if(isPasswordMatching === true) {
-                // Create Token
-                let secret:any = process.env.TOKEN_SECRET;
-                const token = Jwt.sign({
-                    id: user.userId,
-                    email: user.email
-                }, secret, {
-                    expiresIn: 60 * 60 * 24 // expires in 24 hours
-                })
-    
-                return res.header('auth-token', token).json({
-                    user: email, token
-                });
-            }
-        }
-        return res.json({msg: 'User not found!'});
+        const results = await AuthService.login(email, password);
+        return res.json(results);
     } catch (err) {
         return next(Boom.badRequest(err.message));
     }
@@ -58,27 +30,6 @@ export const login = async (req: Request, res: Response , next: NextFunction) =>
 export const verifyToken = async (req: Request, res: Response , next: NextFunction) => {
     // check header or url parameters or post parameters for token
     var token:any = req.query.token;
-    console.log(token);
-    if (!token) {
-      return res.status(400).json({
-        error: true,
-        message: "Token is required."
-      });
-    }
-    
-    // check token that was passed by decoding token using secret
-    let secret:any = process.env.TOKEN_SECRET;
-    Jwt.verify(token, secret, function (err:any, user:any) {
-        if (err) return res.status(401).json({
-            error: true,
-            message: "Invalid token."
-        });
-
-        // get basic user details
-        var userObj = {
-            userId: user.id,
-            email: user.email
-          };
-        return res.json({ user: userObj, token });
-    });
+    const results = AuthService.verify(token);
+    return res.json(results);
 }
